requirements.txt
faker==38.2.0
python-docx==1.2.0
fpdf2==2.8.5
google-generativeai==0.8.5
python-dotenv==1.2.1
pypdf==6.4.2
reportlab==4.4.4
PyPDF2==3.0.1
